Lyapunov Error for Duffing Oscillator
This code calculates the Lyapunov Error (LE) for the Duffing oscillator system, but it's easily usable for other different systems by simply changing the definition of the functions "dq" and "dp". It uses a RK-4th scheme to evaluate the orbits, it grid the phase space and for each point of the grid it calculates the LE. It generates into the executable directory an "orbit.dat" file that has 3 columns such that: (q_points, p_points, LE).
In this repository there is the analogous in python, but it was way too slow to be used, so I wrote a C++ version that is something like 10^5 time faster.